requirements.txt
Flask
twilio
python-dotenv
PyYAML
openai
gunicorn
